+++ amundsen_application/static/js/ducks/user/api/v0.ts
@@ -1,13 +1,14 @@
-import axios, { AxiosResponse, AxiosError } from 'axios';
+import axios, { AxiosResponse } from 'axios';
 
-import { LoggedInUserResponse, UserResponse } from '../types';
+import { LoggedInUser, PeopleUser } from 'interfaces';
+
+export type LoggedInUserResponse = { user: LoggedInUser; msg: string; };
+export type UserResponse = { user: PeopleUser; msg: string; };
 
 export function getLoggedInUser() {
   return axios.get(`/api/auth_user`)
     .then((response: AxiosResponse<LoggedInUserResponse>) => {
       return response.data.user;
-    }).catch((error: AxiosError) => {
-      return {};
     });
 }
 
@@ -15,8 +16,5 @@ export function getUserById(userId: string) {
   return axios.get(`/api/metadata/v0/user?user_id=${userId}`)
   .then((response: AxiosResponse<UserResponse>) => {
     return response.data.user;
-  })
-  .catch((error: AxiosError) => {
-    return {};
   });
 }
